Change coefficient storage to use single instead of double precision. This marginally (about 5%) improves the speed of ndsplineeval(), while halving disk and memory usage. Loss in precision is at the 10^-7 level, which is well beyond the accuracy of the fit.
NB: Old tables still work with the new code, and vice versa.


git-svn-id: http://code.icecube.wisc.edu/svn/projects/photospline/trunk@67084 16731396-06f5-0310-8873-f7f720988828

+++ public/photospline/splinetable.h
@@ -12,7 +12,7 @@ struct splinetable {
 
 	double *periods;
 
-	double *coefficients;
+	float *coefficients;
 	long *naxes;
 	unsigned long *strides;
